:passport_control: created posts delete function on frontend side and…

+++ backend/src/controllers/post.controller.ts
@@ -156,8 +156,35 @@ export const deletePost = async (req: Request, res: Response): Promise<Response>
   const postId: number = parseInt(req.params.id, 10)
 
   try {
-    const result = await postModel.deletePost(postId)
-    return res.status(200).json({ success: `Post deleted!`, result })
+    const jwtPassword = process.env.JWT_PASSWORD
+    const token = req.cookies.token
+
+    if (!token) {
+      return handleResponse(res, 401, 'Unauthorized')
+    }
+
+    if (!jwtPassword) {
+      throw new Error('JWT_PASSWORD is not defined in your .env file')
+    }
+
+    const decodedToken = jwt.verify(token, jwtPassword) as JwtPayload
+
+    // Obtenez le post en utilisant postId
+    const post: Post | null = await postModel.findPostById(postId)
+
+    if (!post) {
+      return handleResponse(res, 404, 'Post not found')
+    }
+
+    // Vérifiez si le roleCode est "4004" ou "2013" ou si l'userId correspond au post
+    if (decodedToken.roleCode === '4004' || decodedToken.roleCode === '2013' || decodedToken.userId === post.userId) {
+      // L'utilisateur a les autorisations nécessaires, vous pouvez maintenant supprimer le post
+      const result = await postModel.deletePost(postId)
+      return res.status(200).json({ success: `Post deleted!`, result })
+    } else {
+      // L'utilisateur n'a pas les autorisations nécessaires
+      return handleResponse(res, 403, 'Forbidden')
+    }
   } catch (err) {
     console.error(err)
     return handleUnexpectedError(res)
